+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## next
+## 0.10.0
 - Add `icy_sixel` feature that uses a Rust implementation of Sixel
 - Remove `lazy_static` dependency in favor of `std::sync::LazyLock`
 - MSRV is now 1.80